Add helm chart
Helm (https://helm.sh) is a templating solution for Kubernetes that
allows us to write one configuration with variables representing our
different environments (local, staging, production). This makes the
config file format *even more* unreadable, but if we're smart the only
things we'll need to edit are the values in `values.yaml`.