wiki: Fix SMW complaining about update key
When Semantic Mediawiki is installed, it generates an upgrade key. If
this key changes, it's a sign that database schema migrations are
needed. To check if it has changed, a file is saved to disk with the
current key.

Since our disks are ephemeral on the pods, every time the wiki is
redeployed (or restarted for any other reason) the upgrade key is
missing and migrations need to be re-run.

This adds the upgrade key in the expected place, so we should be able to
restart until migrations are needed for real.